refactor: ProductType을 page로 이동 #106
- 재사용할 ProductType을 SlidableContainer에서 Page로 옮겼습니다.
